TBLDN-5580-3.2-Display the Price of a Pizza

+++ packages/server/src/application/schema/pizza.schema.ts
@@ -8,7 +8,7 @@ const typeDefs = gql`
     toppingIds: [String!]!
     imgSrc: String!
     toppings: [Topping!]!
-    priceCents: Int!
+    priceCents: Float!
   }
   type Query {
     pizzas: [Pizza!]!
